[Cleanup] Remove rpc handler file, contracts and all contract related data structures for VoiceGuidance. (#843)
* [Cleanup] Remove rpc handler file, contracts and all contract related data structures for VoiceGuidance.
* fix: Fixed format issues.
