+++ core/main/src/firebolt/mod.rs
@@ -36,7 +36,6 @@ pub mod handlers {
     pub mod provider_registrar;
     pub mod second_screen_rpc;
     pub mod user_grants_rpc;
-    pub mod voice_guidance_rpc;
     pub mod wifi_rpc;
 }
 pub mod firebolt_gatekeeper;
