+++ device/thunder_ripple_sdk/src/processors/events/thunder_event_handlers.rs
@@ -15,17 +15,6 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-use std::collections::HashMap;
-
-use ripple_sdk::api::{
-    apps::{AppEvent, AppEventRequest},
-    device::{
-        device_accessibility_data::VoiceGuidanceSettings,
-        device_events::VOICE_GUIDANCE_SETTINGS_CHANGED, device_request::VoiceGuidanceState,
-    },
-};
-use ripple_sdk::serde_json;
-
 use crate::{
     client::thunder_plugin::ThunderPlugin,
     events::thunder_event_processor::{
@@ -36,13 +25,12 @@ use crate::{
             device_events::{DeviceEventCallback, HDCP_CHANGED_EVENT},
             device_request::{AudioProfile, HdcpProfile},
         },
-        extn::extn_client_message::ExtnEvent,
         log::debug,
         tokio,
-        utils::error::RippleError,
     },
     thunder_state::ThunderState,
 };
+use std::collections::HashMap;
 
 use super::super::thunder_device_info::ThunderDeviceInfoRequestProcessor;
 
@@ -108,99 +96,6 @@ impl ThunderEventHandlerProvider for HDCPEventHandler {
 
     fn module() -> String {
         ThunderPlugin::DisplaySettings.callsign_string()
-    }
-}
-
-// -----------------------
-// VoiceGuidance Changed
-
-pub struct VoiceGuidanceEnabledChangedEventHandler;
-
-impl VoiceGuidanceEnabledChangedEventHandler {
-    pub fn handle(
-        state: ThunderState,
-        value: ThunderEventMessage,
-        callback_type: DeviceEventCallback,
-    ) {
-        if let ThunderEventMessage::VoiceGuidance(voice_guidance_state) = value {
-            if let Ok(v) = Self::get_extn_event(voice_guidance_state.clone(), callback_type) {
-                let thunder_state = state.clone();
-                let enabled = voice_guidance_state.state;
-                tokio::spawn(async move {
-                    if let Ok(speed) = ThunderDeviceInfoRequestProcessor::get_voice_guidance_speed(
-                        thunder_state.clone(),
-                    )
-                    .await
-                    {
-                        let settings = VoiceGuidanceSettings { enabled, speed };
-                        let event = ExtnEvent::AppEvent(AppEventRequest::Emit(AppEvent {
-                            event_name: VOICE_GUIDANCE_SETTINGS_CHANGED.to_string(),
-                            result: serde_json::to_value(settings).unwrap(),
-                            context: None,
-                            app_id: None,
-                        }));
-
-                        ThunderEventHandler::callback_device_event(
-                            thunder_state,
-                            VOICE_GUIDANCE_SETTINGS_CHANGED.to_string(),
-                            event,
-                        );
-                    }
-                });
-                ThunderEventHandler::callback_device_event(state, Self::get_mapped_event(), v)
-            }
-        }
-    }
-
-    pub fn is_valid(value: ThunderEventMessage) -> bool {
-        if let ThunderEventMessage::VoiceGuidance(_) = value {
-            return true;
-        }
-        false
-    }
-}
-
-impl ThunderEventHandlerProvider for VoiceGuidanceEnabledChangedEventHandler {
-    type EVENT = VoiceGuidanceState;
-    fn provide(id: String, callback_type: DeviceEventCallback) -> ThunderEventHandler {
-        ThunderEventHandler {
-            request: Self::get_device_request(),
-            handle: Self::handle,
-            is_valid: Self::is_valid,
-            listeners: vec![id],
-            id: Self::get_mapped_event(),
-            callback_type,
-        }
-    }
-
-    fn event_name() -> String {
-        "onVoiceGuidanceChanged".into()
-    }
-
-    fn get_mapped_event() -> String {
-        "voiceguidance.onEnabledChanged".into()
-    }
-
-    fn module() -> String {
-        ThunderPlugin::UserSettings.callsign_string()
-    }
-
-    fn get_extn_event(
-        r: Self::EVENT,
-        callback_type: DeviceEventCallback,
-    ) -> Result<ExtnEvent, RippleError> {
-        let result = serde_json::to_value(r).unwrap();
-        match callback_type {
-            DeviceEventCallback::FireboltAppEvent(_) => {
-                Ok(ExtnEvent::AppEvent(AppEventRequest::Emit(AppEvent {
-                    event_name: Self::get_mapped_event(),
-                    context: None,
-                    result,
-                    app_id: None,
-                })))
-            }
-            DeviceEventCallback::ExtnEvent => Ok(ExtnEvent::Value(result)),
-        }
     }
 }
